requirements.txt
aiogram==2.22.2
python-dotenv==0.21.0
googletrans==3.1.0a0
pydantic==1.10.2
